refactor: dissolve shared/ into drawer-tabs; relocate sticky-action-bar
Move drawer-tab form files to editor/tabs/architecture/drawer-tabs/ with
renames (stack-resource-*-tab → *-tab). Relocate volume-status-variant.test
to lib/tests/ and sticky-action-bar to src/components/ per CLAUDE-only
convention. Update all import paths and relative references. shared/
directory deleted.
