+++ frontend/src/components/sticky-action-bar.tsx
