+++ frontend/src/components/tests/sticky-action-bar.test.tsx
