+++ frontend/src/pages/stacks/components/editor/tabs/architecture/drawer-tabs/dirty-field.tsx
