+++ frontend/src/pages/stacks/components/editor/tabs/architecture/drawer-tabs/env-row.tsx
