+++ frontend/src/pages/stacks/components/editor/tabs/architecture/drawer-tabs/environment-tab.tsx
