+++ frontend/src/pages/stacks/components/editor/tabs/architecture/drawer-tabs/tests/env-row.test.tsx
