+++ frontend/src/pages/stacks/components/editor/tabs/architecture/drawer-tabs/tests/use-resource-tab-props.test.tsx
